requirements.txt
discord.py
feedparser